refactor(*): use "one class=one file" naming convention, update readme
