+++ back/flag-service/src/flag/errors/CooldownTimerHasNotEndedYetError.ts
@@ -1,0 +1,1 @@
+export class CooldownTimerHasNotEndedYetError extends Error {}
