+++ back/flag-service/src/flag/errors/UserAlreadyOwnAPixelError.ts
@@ -1,0 +1,1 @@
+export class UserAlreadyOwnAPixelError extends Error {}
