--- back/flag-service/src/flag/flag-errors.ts
@@ -1,3 +1,0 @@
-export class UserAlreadyOwnAPixelError extends Error {}
-
-export class CooldownTimerHasNotEndedYet extends Error {}
